+++ lib/drives/pages/create_drive_page.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:rrule/rrule.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
 import '../../account/models/profile.dart';
@@ -17,6 +16,8 @@ import '../../util/trip/trip.dart';
 import '../models/drive.dart';
 import '../models/recurring_drive.dart';
 import '../pages/drive_detail_page.dart';
+import '../util/recurrence.dart';
+import '../util/week_day.dart';
 
 class CreateDrivePage extends StatefulWidget {
   const CreateDrivePage({super.key});
@@ -310,7 +311,7 @@ class _CreateDriveFormState extends State<CreateDriveForm> {
   }
 
   Widget buildWeekDayPicker() {
-    return WeekDaySelector(
+    return WeekDayPicker(
       weekDays: _recurrenceOptions.weekDays,
       onWeekDaysChanged: (List<WeekDay> weekDays) => setState(() {
         _recurrenceOptions.weekDays = weekDays;
@@ -373,11 +374,13 @@ class _CreateDriveFormState extends State<CreateDriveForm> {
   }
 
   Widget buildUntilPicker() {
-    return InkWell(
-      onTap: showRecurrenceEndDialog,
-      child: Padding(
-        padding: const EdgeInsets.symmetric(vertical: 12),
-        child: Text(_recurrenceOptions.getEndChoiceName(context)),
+    return SizedBox(
+      width: 200,
+      child: TextFormField(
+        decoration: const InputDecoration(border: OutlineInputBorder()),
+        onTap: showRecurrenceEndDialog,
+        readOnly: true,
+        controller: _recurrenceOptions.endChoiceController,
       ),
     );
   }
@@ -390,7 +393,7 @@ class _CreateDriveFormState extends State<CreateDriveForm> {
         builder: (BuildContext context, void Function(VoidCallback) innerSetState) {
           void onChanged(RecurrenceEndChoice? value) {
             innerSetState(() {
-              _recurrenceOptions.endChoice = value!;
+              _recurrenceOptions.setEndChoice(value!, context);
             });
           }
 
@@ -437,7 +440,7 @@ class _CreateDriveFormState extends State<CreateDriveForm> {
                             ).then((DateTime? value) {
                               if (value != null) {
                                 innerSetState(() {
-                                  _recurrenceOptions.setCustomDate(value);
+                                  _recurrenceOptions.setCustomDate(value, context);
                                 });
                               }
                             }),
@@ -467,6 +470,7 @@ class _CreateDriveFormState extends State<CreateDriveForm> {
                             onChanged: (String value) {
                               innerSetState(() {
                                 _recurrenceOptions.customEndIntervalChoice.intervalSize = int.tryParse(value);
+                                _recurrenceOptions.rebuildEndChoiceController(context);
                               });
                             },
                           );
@@ -525,6 +529,7 @@ class _CreateDriveFormState extends State<CreateDriveForm> {
                             onChanged: (String value) {
                               innerSetState(() {
                                 _recurrenceOptions.customEndOccurrenceChoice.occurrences = int.tryParse(value);
+                                _recurrenceOptions.rebuildEndChoiceController(context);
                               });
                             },
                           );
@@ -580,401 +585,3 @@ class _CreateDriveFormState extends State<CreateDriveForm> {
     );
   }
 }
-
-class WeekDaySelector extends FormField<List<WeekDay>> {
-  final List<WeekDay> weekDays;
-  final ValueChanged<List<WeekDay>> onWeekDaysChanged;
-
-  WeekDaySelector({
-    super.key,
-    this.weekDays = const <WeekDay>[],
-    required this.onWeekDaysChanged,
-  }) : super(
-          builder: (FormFieldState<List<WeekDay>> state) {
-            return Column(
-              children: <Widget>[
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                  children: WeekDay.values
-                      .map(
-                        (WeekDay weekDay) => WeekDayButton(
-                          weekDay: weekDay,
-                          selected: weekDays.contains(weekDay),
-                          onPressed: () {
-                            if (weekDays.contains(weekDay)) {
-                              weekDays.remove(weekDay);
-                            } else {
-                              weekDays.add(weekDay);
-                            }
-                            state.didChange(weekDays);
-                          },
-                        ),
-                      )
-                      .toList(),
-                ),
-                if (state.hasError)
-                  Padding(
-                    padding: const EdgeInsets.only(top: 8.0),
-                    child: Text(
-                      state.errorText!,
-                      style: const TextStyle(color: Colors.red),
-                    ),
-                  ),
-              ],
-            );
-          },
-          validator: (List<WeekDay>? value) {
-            if (value == null || value.isEmpty) {
-              return 'Please select at least one day';
-            }
-            return null;
-          },
-        );
-}
-
-class WeekDayButton extends StatelessWidget {
-  const WeekDayButton({
-    super.key,
-    required this.weekDay,
-    required this.onPressed,
-    required this.selected,
-  });
-
-  final WeekDay weekDay;
-  final VoidCallback onPressed;
-  final bool selected;
-
-  @override
-  Widget build(BuildContext context) {
-    return Semantics(
-      button: true,
-      child: ElevatedButton(
-        style: ElevatedButton.styleFrom(
-          shape: const CircleBorder(),
-          foregroundColor: Colors.white,
-          minimumSize: const Size.fromRadius(15),
-          backgroundColor: selected ? Theme.of(context).colorScheme.primary : Colors.grey,
-        ),
-        onPressed: onPressed,
-        child: Text(weekDay.getAbbreviation(context)),
-      ),
-    );
-  }
-}
-
-class RecurrenceOptions {
-  bool enabled = false;
-
-  late RecurrenceEndChoice _endChoice;
-
-  RecurrenceEndChoice get endChoice => _endChoice;
-
-  set endChoice(RecurrenceEndChoice value) {
-    _endChoice = value;
-    if (value.isCustom) {
-      if (value is RecurrenceEndChoiceDate) {
-        customEndDateChoice = value;
-      } else if (value is RecurrenceEndChoiceInterval) {
-        customEndIntervalChoice = value;
-      } else if (value is RecurrenceEndChoiceOccurrence) {
-        customEndOccurrenceChoice = value;
-      }
-    }
-  }
-
-  late final List<WeekDay> weekDays;
-
-  late RecurrenceInterval recurrenceInterval;
-  final TextEditingController recurrenceIntervalSizeController = TextEditingController();
-  final TextEditingController recurrenceIntervalTypeController = TextEditingController();
-
-  void setRecurrenceIntervalType(RecurrenceIntervalType type, BuildContext context) {
-    recurrenceInterval.intervalType = type;
-    recurrenceIntervalTypeController.text = type.getName(context);
-  }
-
-  RecurrenceEndChoiceDate customEndDateChoice = RecurrenceEndChoiceDate(null, isCustom: true);
-  final TextEditingController customEndDateController = TextEditingController();
-
-  void setCustomDate(DateTime date) {
-    customEndDateChoice.date = date;
-    customEndDateController.text = localeManager.formatDate(date);
-  }
-
-  RecurrenceEndChoiceInterval customEndIntervalChoice = RecurrenceEndChoiceInterval(null, null, isCustom: true);
-  final TextEditingController customEndIntervalSizeController = TextEditingController();
-  final TextEditingController customEndIntervalTypeController = TextEditingController();
-
-  void setCustomEndIntervalType(RecurrenceIntervalType type, BuildContext context) {
-    customEndIntervalChoice.intervalType = type;
-    customEndIntervalTypeController.text = type.getName(context);
-  }
-
-  RecurrenceEndChoiceOccurrence customEndOccurrenceChoice = RecurrenceEndChoiceOccurrence(null, isCustom: true);
-  final TextEditingController customEndOccurrenceController = TextEditingController();
-
-  void setCustomEndOccurrence(int occurrence) {
-    customEndOccurrenceChoice.occurrences = occurrence;
-    customEndOccurrenceController.text = occurrence.toString();
-  }
-
-  String? validationError;
-
-  RecurrenceEndChoice getRecurrenceEndChoice(RecurrenceEndType type) {
-    switch (type) {
-      case RecurrenceEndType.date:
-        return customEndDateChoice;
-      case RecurrenceEndType.interval:
-        return customEndIntervalChoice;
-      case RecurrenceEndType.occurrence:
-        return customEndOccurrenceChoice;
-    }
-  }
-
-  String getEndChoiceName(BuildContext context) => endChoice.getName(context, weekDays: weekDays);
-
-  RecurrenceOptions({
-    required RecurrenceEndChoice endChoice,
-    required this.recurrenceInterval,
-    List<WeekDay>? weekDays,
-    required BuildContext context,
-  }) {
-    _endChoice = endChoice;
-    recurrenceIntervalSizeController.text = recurrenceInterval.intervalSize.toString();
-    recurrenceIntervalTypeController.text = recurrenceInterval.intervalType.getName(context);
-    this.weekDays = weekDays ?? <WeekDay>[];
-  }
-
-  void dispose() {
-    recurrenceIntervalSizeController.dispose();
-    recurrenceIntervalTypeController.dispose();
-    customEndDateController.dispose();
-    customEndIntervalSizeController.dispose();
-    customEndIntervalTypeController.dispose();
-    customEndOccurrenceController.dispose();
-  }
-
-  bool validate({bool createError = true}) {
-    final String? error = endChoice.validate(weekDays: weekDays);
-    if (createError || error == null) {
-      validationError = error;
-    }
-    return error == null;
-  }
-
-  RecurrenceRule get recurrenceRule {
-    final Frequency frequency = recurrenceInterval.intervalType.frequency;
-    final int interval = recurrenceInterval.intervalSize!;
-    final Set<ByWeekDayEntry> byWeekDays = weekDays.map((WeekDay weekDay) => ByWeekDayEntry(weekDay.index + 1)).toSet();
-
-    DateTime? until;
-
-    if (_endChoice.type == RecurrenceEndType.date) {
-      until = (endChoice as RecurrenceEndChoiceDate).date;
-    } else if (_endChoice.type == RecurrenceEndType.interval) {
-      until = (endChoice as RecurrenceEndChoiceInterval).date;
-    } else if (_endChoice.type == RecurrenceEndType.occurrence) {
-      return RecurrenceRule(
-        frequency: frequency,
-        interval: interval,
-        byWeekDays: byWeekDays,
-        count: (endChoice as RecurrenceEndChoiceOccurrence).occurrences,
-      );
-    }
-
-    return RecurrenceRule(
-      frequency: frequency,
-      interval: interval,
-      byWeekDays: byWeekDays,
-      until: until!.copyWith(hour: 23, minute: 59).toUtc(),
-    );
-  }
-}
-
-class RecurrenceInterval {
-  int? intervalSize;
-  RecurrenceIntervalType intervalType;
-
-  RecurrenceInterval(this.intervalSize, this.intervalType);
-
-  String getName(BuildContext context, {List<WeekDay>? weekDays}) {
-    final String? validationError = validate();
-    if (validationError != null) return validationError;
-
-    switch (intervalType) {
-      case RecurrenceIntervalType.days:
-        return 'Every $intervalSize ${intervalSize == 1 ? 'day' : 'days'}';
-      case RecurrenceIntervalType.weeks:
-        return 'Every $intervalSize ${intervalSize == 1 ? 'week' : 'weeks'}';
-      case RecurrenceIntervalType.months:
-        return 'Every $intervalSize ${intervalSize == 1 ? 'month' : 'months'}';
-      case RecurrenceIntervalType.years:
-        return 'Every $intervalSize ${intervalSize == 1 ? 'year' : 'years'}';
-    }
-  }
-
-  String? validate() => intervalSize == null ? 'Interval size needed' : null;
-}
-
-abstract class RecurrenceEndChoice {
-  RecurrenceEndType type;
-  final bool isCustom;
-
-  RecurrenceEndChoice({this.type = RecurrenceEndType.occurrence, this.isCustom = false});
-
-  String getName(BuildContext context, {List<WeekDay>? weekDays});
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) || other is RecurrenceEndChoice && type == other.type && isCustom == other.isCustom;
-
-  @override
-  int get hashCode => type.hashCode ^ isCustom.hashCode;
-
-  String? validate({List<WeekDay>? weekDays});
-}
-
-class RecurrenceEndChoiceDate extends RecurrenceEndChoice {
-  DateTime? date;
-
-  RecurrenceEndChoiceDate(this.date, {super.isCustom})
-      : assert(date != null || isCustom),
-        super(type: RecurrenceEndType.date);
-
-  @override
-  String getName(BuildContext context, {List<WeekDay>? weekDays}) =>
-      validate(weekDays: weekDays) ?? 'Until ${localeManager.formatDate(date!)}';
-
-  @override
-  bool operator ==(Object other) =>
-      other is RecurrenceEndChoiceDate && isCustom == other.isCustom && date == other.date;
-
-  @override
-  int get hashCode => isCustom.hashCode ^ date.hashCode;
-
-  @override
-  String? validate({List<WeekDay>? weekDays}) => date == null
-      ? 'Date needed'
-      : date!.isAfter(DateTime.now().add(const Duration(days: 10 * 365)))
-          ? 'Date too far in the future'
-          : null;
-}
-
-class RecurrenceEndChoiceInterval extends RecurrenceEndChoice {
-  int? intervalSize;
-  RecurrenceIntervalType? intervalType;
-
-  RecurrenceEndChoiceInterval(this.intervalSize, this.intervalType, {super.isCustom})
-      : assert((intervalSize != null && intervalType != null) || isCustom),
-        super(type: RecurrenceEndType.interval);
-
-  DateTime get date {
-    final DateTime now = DateTime.now();
-    switch (intervalType!) {
-      case RecurrenceIntervalType.days:
-        return now.add(Duration(days: intervalSize!));
-      case RecurrenceIntervalType.weeks:
-        return now.add(Duration(days: intervalSize! * 7));
-      case RecurrenceIntervalType.months:
-        final int nextMonth = now.month + intervalSize!;
-        return DateTime(now.year + (nextMonth + 1) ~/ 12, (nextMonth + 1) % 12 - 1, now.day);
-      case RecurrenceIntervalType.years:
-        return DateTime(now.year + intervalSize!, now.month, now.day);
-    }
-  }
-
-  @override
-  String getName(BuildContext context, {List<WeekDay>? weekDays}) {
-    final String? validationError = validate(weekDays: weekDays);
-    if (validationError != null) return validationError;
-
-    switch (intervalType!) {
-      case RecurrenceIntervalType.days:
-        return 'For $intervalSize ${intervalSize == 1 ? 'day' : 'days'}';
-      case RecurrenceIntervalType.weeks:
-        return 'For $intervalSize ${intervalSize == 1 ? 'week' : 'weeks'}';
-      case RecurrenceIntervalType.months:
-        return 'For $intervalSize ${intervalSize == 1 ? 'month' : 'months'}';
-      case RecurrenceIntervalType.years:
-        return 'For $intervalSize ${intervalSize == 1 ? 'year' : 'years'}';
-    }
-  }
-
-  @override
-  String? validate({List<WeekDay>? weekDays}) {
-    if (intervalSize == null) return 'Interval size needed';
-    if (intervalType == null) return 'Interval type needed';
-    if (intervalSize! >= 100) return 'Interval too large';
-    if (intervalSize! <= 0) return 'Interval is negative';
-    if (intervalType == RecurrenceIntervalType.years && intervalSize! > 10) return 'Interval too large';
-    return null;
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      other is RecurrenceEndChoiceInterval &&
-      isCustom == other.isCustom &&
-      intervalSize == other.intervalSize &&
-      intervalType == other.intervalType;
-
-  @override
-  int get hashCode => isCustom.hashCode ^ intervalSize.hashCode ^ intervalType.hashCode;
-}
-
-enum RecurrenceIntervalType { days, weeks, months, years }
-
-extension RecurrenceEndIntervalTypeExtension on RecurrenceIntervalType {
-  String getName(BuildContext context) {
-    switch (this) {
-      case RecurrenceIntervalType.days:
-        return 'Days';
-      case RecurrenceIntervalType.weeks:
-        return 'Weeks';
-      case RecurrenceIntervalType.months:
-        return 'Months';
-      case RecurrenceIntervalType.years:
-        return 'Years';
-    }
-  }
-
-  Frequency get frequency {
-    switch (this) {
-      case RecurrenceIntervalType.days:
-        return Frequency.daily;
-      case RecurrenceIntervalType.weeks:
-        return Frequency.weekly;
-      case RecurrenceIntervalType.months:
-        return Frequency.monthly;
-      case RecurrenceIntervalType.years:
-        return Frequency.yearly;
-    }
-  }
-}
-
-class RecurrenceEndChoiceOccurrence extends RecurrenceEndChoice {
-  int? occurrences;
-
-  RecurrenceEndChoiceOccurrence(this.occurrences, {super.isCustom})
-      : assert(occurrences != null || isCustom),
-        super(type: RecurrenceEndType.occurrence);
-
-  @override
-  String getName(BuildContext context, {List<WeekDay>? weekDays}) =>
-      validate(weekDays: weekDays) ?? 'After $occurrences ${occurrences == 1 ? 'occurrence' : 'occurrences'}';
-
-  @override
-  String? validate({List<WeekDay>? weekDays}) => occurrences == null
-      ? 'Occurrences needed'
-      : occurrences! >= 100
-          ? 'Too many occurrences'
-          : null;
-
-  @override
-  bool operator ==(Object other) =>
-      other is RecurrenceEndChoiceOccurrence && isCustom == other.isCustom && occurrences == other.occurrences;
-
-  @override
-  int get hashCode => isCustom.hashCode ^ occurrences.hashCode;
-}
-
-enum RecurrenceEndType { date, interval, occurrence }
